requirements.txt
# requirements.txt
fastapi
uvicorn
gunicorn
python-multipart
streamlit
requests
openai
python-dotenv
pdfplumber
openpyxl
python-docx
pandas
pillow
pytesseract
